draft implementation(7)

+++ src/domains/devices.service.ts
@@ -15,13 +15,11 @@ export const devicesService = {
       if (!result?.userId) return HTTP_STATUSES.UNAUTHORIZED_401
 
       const user: UserType | null = await usersQueryRepository.fetchAllUserData(result.userId)
-      if (!user) return HTTP_STATUSES.UNAUTHORIZED_401
-      if (user.invalidRefreshTokens.includes(req.cookies.refreshToken)) return HTTP_STATUSES.UNAUTHORIZED_401
+      if (!user || user.invalidRefreshTokens.includes(req.cookies.refreshToken)) return HTTP_STATUSES.UNAUTHORIZED_401
 
       const isDeviceIdExist = await refreshTokenDevicesRepository.findDeviceId(deviceId)
       if (!isDeviceIdExist) return HTTP_STATUSES.NOT_FOUND_404
-      if (user.id !== result.userId) return HTTP_STATUSES.FORBIDDEN_403
-      if (result.deviceId === deviceId) return HTTP_STATUSES.NOT_FOUND_404
+      if (user.id !== result.userId || result.deviceId === deviceId) return HTTP_STATUSES.FORBIDDEN_403
 
       const isDeleted = await refreshTokenDevicesRepository.removeSpecifiedSession(result.userId, deviceId)
